Move syntax highlighting to its own module.
